veux: Build sensorservice 1.0 HAL
Fixes:
W HidlServiceManagement: Waited one second for android.frameworks.sensorservice@1.0::ISensorManager/default
I hwservicemanager: Since android.frameworks.sensorservice@1.0::ISensorManager/default is not registered, trying to start it as a lazy HAL (if it's not configured to be a lazy HAL, it may be stuck starting or still starting).
I HidlServiceManagement: getService: Trying again for android.frameworks.sensorservice@1.0::ISensorManager/default...
W libc    : Unable to set property "ctl.interface_start" to "android.frameworks.sensorservice@1.0::ISensorManager/default": error code: 0x20
I hwservicemanager: Tried to start android.frameworks.sensorservice@1.0::ISensorManager/default as a lazy service, but was unable to. Usually this happens when a service is not installed, but if the service is intended to be used as a lazy service, then it may be configured incorrectly.

Change-Id: If6fdef9dc39daaba9138fe3f8acf06cb6a0e78f8
Signed-off-by: Frost <159105703+Karan-Frost@users.noreply.github.com>